requirements.txt
absl-py==1.4.0
appdirs==1.4.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.6
cycler==0.11.0
docker-pycreds==0.4.0
gitdb==4.0.10
GitPython==3.1.32
glfw==2.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.3
imgui==1.3.0
importlib-metadata==6.8.0
kiwisolver==1.4.4
lazy_loader==0.3
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.4.2
mrcfile==1.4.3
networkx==3.1
ninja==1.10.2
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
pathtools==0.1.2
Pillow==10.0.0
plyfile==1.0.1
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyOpenGL==3.1.5
pyparsing==3.1.0
pyspng==0.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tifffile==2023.7.10
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.62.2
trimesh==3.22.5
typing_extensions==4.7.1
urllib3==1.26.16
wandb==0.15.5
Werkzeug==2.3.6
zipp==3.16.2
